backend_layer_a/requirements.txt
fastapi
uvicorn
langgraph
langchain
pydantic
python-multipart
httpx
requests
pypdf
langchain-google-genai
google-generativeai
python-dotenv
